put statusline themes hl groups in one file & fix https://github.com/NvChad/ui/issues/43
